+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Change Log
 
 * Creation of `Widgets` without a title but with a reference (or `ref`) now supported via the `show_title_value` keyword.
 * Creation of `Activity` now uses parent's `classification` if provided.
+* Added `page_size` input the `add_scope_widget` method of the WidgetManager class.
 * :+1: dependent versions for development: sphinx (2.4.2), pre-commit (2.1.0), requests (2.23.0)
 
 3.4.0 (17FEB20)